+++ src/Data/Entity/CustomEntity/CustomEntityEntity.php
@@ -1,0 +1,18 @@
+<?php declare(strict_types=1);
+namespace Vin\ShopwareSdk\Data\Entity\CustomEntity;
+
+use Vin\ShopwareSdk\Data\Entity\Entity;
+
+/**
+ * Shopware Entity Mapping Class
+ *
+ * This class is generated dynamically following SW entities schema
+ */
+class CustomEntityEntity extends Entity
+{
+    public ?string $name = null;
+
+    public ?array $fields = null;
+
+    public ?string $appId = null;
+}